m12n: move descriptor-related classes to oak-base
