+++ oak-base/src/main/java/org/apache/jackrabbit/oak/spi/descriptors/AggregatingDescriptors.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.oak.commons.jmx;
+package org.apache.jackrabbit.oak.spi.descriptors;
 
 import java.util.Collections;
 import java.util.HashSet;
